feat(dashboard): add cursor styling and refactor footer constants
- Add cursor-pointer to language selector and theme toggle
- Use COMPANY_OWNER constant in footer copyright
